fix(cli): move discovery init dir
